requirements.txt
pyyaml
reportlab
qrcode[pil]
svglib